tests/requirements.txt
tox
pytest
pytest-flask
pytest-cov
coverage
flake8
